requirements.txt
gradio
torch
torchvision
albumentations
pillow
timm
numpy<2